Moving BehaviorFormContext to the contexts folder (#594)
* Moving context files
