+++ src/Components/ADT3DSceneBuilder/ADT3DSceneBuilder.tsx
@@ -134,8 +134,6 @@ const ADT3DSceneBuilderBase: React.FC<IADT3DSceneBuilderCardProps> = (
     const classNames = getClassNames(styles, { theme: fluentTheme });
 
     // state
-    const [behaviorToEdit, setBehaviorToEdit] = useState<IBehavior>(null);
-
     const previouslyColoredMeshItems = useRef([]);
     const elementContextualMenuItems = useRef([]);
     const behaviorContextualMenuItems = useRef<IContextualMenuItem[]>([]);
@@ -528,7 +526,6 @@ const ADT3DSceneBuilderBase: React.FC<IADT3DSceneBuilderCardProps> = (
                         type: SET_ADT_SCENE_BUILDER_MODE,
                         payload: ADT3DSceneBuilderMode.EditBehavior
                     });
-                    setBehaviorToEdit(behavior);
                 },
                 onMouseOver: () => {
                     // get elements that are contained in the hovered behavior
